Implementing UI and frontend for Parent-Child relationships

Closes #406, #548, #551, and #553

See merge request zesje/zesje!336

+++ zesje/api/feedback.py
@@ -5,6 +5,18 @@ from flask_restful import Resource, reqparse
 from ..database import db, Problem, FeedbackOption, Solution
 
 
+def feedback_to_data(feedback, full_children=True):
+    return {
+        'id': feedback.id,
+        'name': feedback.text,
+        'description': feedback.description,
+        'score': feedback.score,
+        'parent': feedback.parent_id,
+        'used': len(feedback.solutions),
+        'children': [feedback_to_data(child) if full_children else child.id for child in feedback.children]
+    }
+
+
 class Feedback(Resource):
     """ List of feedback options of a problem """
 
@@ -26,20 +38,9 @@ class Feedback(Resource):
         """
 
         if (problem := Problem.query.get(problem_id)) is None:
-            return dict(status=404, message=f"Problem with id #{problem_id} does not exist"), 404
+            return dict(status=422, message=f"Problem with id #{problem_id} does not exist"), 422
 
-        return [
-            {
-                'id': fb.id,
-                'name': fb.text,
-                'description': fb.description,
-                'score': fb.score,
-                'parent': fb.parent_id,
-                'used': len(fb.solutions),
-                'children': [feedback.id for feedback in fb.children]
-            }
-            for fb in FeedbackOption.query.filter(FeedbackOption.problem == problem)
-        ]
+        return feedback_to_data(problem.root_feedback)
 
     post_parser = reqparse.RequestParser()
     post_parser.add_argument('name', type=str, required=True)
@@ -59,22 +60,19 @@ class Feedback(Resource):
         """
 
         if (problem := Problem.query.get(problem_id)) is None:
-            return dict(status=404, message=f"Problem with id #{problem_id} does not exist"), 404
+            return dict(status=422, message=f"Problem with id #{problem_id} does not exist"), 422
 
         args = self.post_parser.parse_args()
-        parent_id = args.parent
-        fb = FeedbackOption(problem=problem, text=args.name, description=args.description, score=args.score)
+        parent = FeedbackOption.query.get(args.parent)
+        if parent is None:
+            return dict(status=422, message=f"FeedbackOption with id #{args.parent} does not exist"), 422
 
-        if parent_id is None:  # Will only be the case when the initial "null" root has to be created
-            # All feedback options must have a parent
-            db.session.add(fb)
-
-        else:
-            parent = FeedbackOption.query.get(parent_id)
-            if parent is None:
-                return dict(status=404, message=f"FeedbackOption with id #{parent_id} does not exist"), 404
-
-            parent.children.append(fb)
+        fb = FeedbackOption(problem=problem,
+                            text=args.name,
+                            description=args.description,
+                            score=args.score,
+                            parent=parent)
+        db.session.add(fb)
 
         db.session.commit()
 
@@ -143,8 +141,11 @@ class Feedback(Resource):
         if problem.id != problem_id:
             return dict(status=400, message="Feedback option does not belong to this problem."), 400
         if fb.mc_option:
-            return dict(status=403, message='Cannot delete feedback option'
-                                            + ' attached to a multiple choice option.'), 403
+            return dict(status=405, message='Cannot delete feedback option'
+                                            + ' attached to a multiple choice option.'), 405
+        if fb.parent_id is None:
+            return dict(status=405, message='Cannot delete root feedback option.'), 405
+
         # All feedback options, that are the child of the original feedback option will be deleted
 
         db.session.delete(fb)
